InteractiveViewBox: Move code to utils
