+++ Orange/widgets/visualize/utils/plotutils.py
@@ -1,10 +1,16 @@
 import numpy as np
 
-from AnyQt.QtWidgets import QGraphicsLineItem, QGraphicsSceneMouseEvent
-from AnyQt.QtCore import QRectF, QLineF, QObject, QEvent
+from AnyQt.QtCore import (
+    QRectF, QLineF, QObject, QEvent, Qt, pyqtSignal as Signal
+)
 from AnyQt.QtGui import QTransform
+from AnyQt.QtWidgets import (
+    QGraphicsLineItem, QGraphicsSceneMouseEvent, QPinchGesture
+)
 
 import pyqtgraph as pg
+
+from Orange.widgets.utils.plot import SELECT, PANNING, ZOOMING
 
 
 class TextItem(pg.TextItem):
@@ -121,3 +127,159 @@ class MouseEventDelegate(HelpEventDelegate):
         if isinstance(ev, QGraphicsSceneMouseEvent):
             self.delegate2(ev)
         return super().eventFilter(obj, ev)
+
+
+class InteractiveViewBox(pg.ViewBox):
+    def __init__(self, graph, enable_menu=False):
+        self.init_history()
+        pg.ViewBox.__init__(self, enableMenu=enable_menu)
+        self.graph = graph
+        self.setMouseMode(self.PanMode)
+        self.grabGesture(Qt.PinchGesture)
+
+    def _dragtip_pos(self):
+        return 10, 10
+
+    def safe_update_scale_box(self, buttonDownPos, currentPos):
+        x, y = currentPos
+        if buttonDownPos[0] == x:
+            x += 1
+        if buttonDownPos[1] == y:
+            y += 1
+        self.updateScaleBox(buttonDownPos, pg.Point(x, y))
+
+    # noinspection PyPep8Naming,PyMethodOverriding
+    def mouseDragEvent(self, ev, axis=None):
+        if self.graph.state == SELECT and axis is None:
+            ev.accept()
+            pos = ev.pos()
+            if ev.button() == Qt.LeftButton:
+                self.safe_update_scale_box(ev.buttonDownPos(), ev.pos())
+                scene = self.scene()
+                dragtip = scene.drag_tooltip
+                if ev.isFinish():
+                    dragtip.hide()
+                    self.rbScaleBox.hide()
+                    pixel_rect = QRectF(ev.buttonDownPos(ev.button()), pos)
+                    value_rect = self.childGroup.mapRectFromParent(pixel_rect)
+                    self.graph.select_by_rectangle(value_rect)
+                else:
+                    dragtip.setPos(*self._dragtip_pos())
+                    dragtip.show()  # although possibly already shown
+                    self.safe_update_scale_box(ev.buttonDownPos(), ev.pos())
+        elif self.graph.state == ZOOMING or self.graph.state == PANNING:
+            ev.ignore()
+            super().mouseDragEvent(ev, axis=axis)
+        else:
+            ev.ignore()
+
+    def updateAutoRange(self):
+        # indirectly called by the autorange button on the graph
+        super().updateAutoRange()
+        self.tag_history()
+
+    def tag_history(self):
+        #add current view to history if it differs from the last view
+        if self.axHistory:
+            currentview = self.viewRect()
+            lastview = self.axHistory[self.axHistoryPointer]
+            inters = currentview & lastview
+            united = currentview.united(lastview)
+            if inters.width()*inters.height()/(united.width()*united.height()) > 0.95:
+                return
+        self.axHistoryPointer += 1
+        self.axHistory = self.axHistory[:self.axHistoryPointer] + \
+                         [self.viewRect()]
+
+    def init_history(self):
+        self.axHistory = []
+        self.axHistoryPointer = -1
+
+    def autoRange(self, padding=None, items=None, item=None):
+        super().autoRange(padding=padding, items=items, item=item)
+        self.tag_history()
+
+    def suggestPadding(self, axis): #no padding so that undo works correcty
+        return 0.
+
+    def scaleHistory(self, d):
+        self.tag_history()
+        super().scaleHistory(d)
+
+    def mouseClickEvent(self, ev):
+        if ev.button() == Qt.RightButton:  # undo zoom
+            self.scaleHistory(-1)
+        else:
+            ev.accept()
+            self.graph.unselect_all()
+
+    def sceneEvent(self, event):
+        if event.type() == QEvent.Gesture:
+            return self.gestureEvent(event)
+        return super().sceneEvent(event)
+
+    def gestureEvent(self, event):
+        gesture = event.gesture(Qt.PinchGesture)
+        if gesture.state() == Qt.GestureStarted:
+            event.accept(gesture)
+        elif gesture.changeFlags() & QPinchGesture.ScaleFactorChanged:
+            center = self.mapSceneToView(gesture.centerPoint())
+            scale_prev = gesture.lastScaleFactor()
+            scale = gesture.scaleFactor()
+            if scale_prev != 0:
+                scale = scale / scale_prev
+            if scale > 0:
+                self.scaleBy((1 / scale, 1 / scale), center)
+        elif gesture.state() == Qt.GestureFinished:
+            self.tag_history()
+
+        return True
+
+
+class VizInteractiveViewBox(InteractiveViewBox):
+    started = Signal()
+    moved = Signal()
+    finished = Signal()
+
+    def __init__(self, graph, enable_menu=False):
+        self.mouse_state = 0
+        self.point_i = None
+        super().__init__(graph, enable_menu)
+
+    def mousePressEvent(self, ev):
+        super().mousePressEvent(ev)
+        pos = self.childGroup.mapFromParent(ev.pos())
+        if self.graph.can_show_indicator(pos)[0]:
+            self.setCursor(Qt.ClosedHandCursor)
+
+    def mouseDragEvent(self, ev, axis=None):
+        pos = self.childGroup.mapFromParent(ev.pos())
+        can_show, point_i = self.graph.can_show_indicator(pos)
+        if ev.button() != Qt.LeftButton or (ev.start and not can_show):
+            self.mouse_state = 2
+        if self.mouse_state == 2:
+            if ev.finish:
+                self.mouse_state = 0
+            super().mouseDragEvent(ev, axis)
+            return
+
+        ev.accept()
+        if ev.start:
+            self.setCursor(Qt.ClosedHandCursor)
+            self.mouse_state = 1
+            self.point_i = point_i
+            self.started.emit()
+
+        if self.mouse_state == 1:
+            self.graph.set_point(self.point_i, pos.x(), pos.y())
+            if ev.finish:
+                self.setCursor(Qt.OpenHandCursor)
+                self.mouse_state = 0
+                self.finished.emit()
+            else:
+                self._show_tooltip(ev)
+                self.moved.emit()
+            self.graph.show_indicator(self.point_i)
+
+    def _show_tooltip(self, ev):
+        pass
